[new_header] reconstruct header files
