+++ include/core.h
@@ -1,18 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   hashmap.h                                          :+:      :+:    :+:   */
+/*   ../../../include/minishell.h                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gimmingyu <gimmingyu@student.42.fr>        +#+  +:+       +#+        */
+/*   By: mingkim <mingkim@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/10/31 10:41:31 by gimmingyu         #+#    #+#             */
-/*   Updated: 2022/10/31 15:10:50 by gimmingyu        ###   ########.fr       */
+/*   Created: 2022/11/01 12:15:15 by mingkim           #+#    #+#             */
+/*   Updated: 2022/11/02 14:44:11 by mingkim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef HASHTABLE_H
-# define HASHTABLE_H
+#ifndef CORE_H
+# define CORE_H
 
-# include "../../../include/minishell.h"
+# include "./struct.h"
+
+/* in ../src/internal/core/ */
+/* prompt.c */
+char			*create_prompt(void);
+/* signal.c */
+void			sig_handler(int signal);
+void			setting_signal(void);
 
 #endif
