+++ src/internal/builtin/builtin.c
@@ -10,4 +10,4 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "builtin.h"
+#include "../../../include/minishell.h"
