+++ src/internal/builtin/cd.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "builtin.h"
+#include "../../../include/minishell.h"
 
 void	_invalid_err_cd(const char *str)
 {
