+++ src/internal/builtin/echo.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "builtin.h"
+#include "../../../include/minishell.h"
 
 int	builtin_echo(const char *str, int fd, int flag)
 {
