+++ src/internal/doublylist/delete.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "doublylist.h"
+#include "../../../include/minishell.h"
 /**
  * 인덱스 0에 접근할 때  || 인덱스 (len -1)에 접근할 때
  *
