--- src/internal/doublylist/doublylist.h
@@ -1,18 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   doublylist.h                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mingkim <mingkim@student.42.fr>            +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/10/14 15:29:32 by mingkim           #+#    #+#             */
-/*   Updated: 2022/10/14 15:33:14 by mingkim          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#ifndef DOUBLYLIST_H
-# define DOUBLYLIST_H
-
-# include "../../../include/minishell.h"
-
-#endif
