--- src/internal/execve/executor.h
@@ -1,18 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   executor.h                                         :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mingkim <mingkim@student.42.fr>            +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/10/14 15:27:24 by mingkim           #+#    #+#             */
-/*   Updated: 2022/10/17 11:18:28 by mingkim          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#ifndef EXECUTOR_H
-# define EXECUTOR_H
-
-# include "../../../include/minishell.h"
-
-#endif
