+++ src/internal/hashtable/create.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "hashtable.h"
+#include "../../../include/minishell.h"
 
 t_hashtable	*create_hashtable(size_t size)
 {
