+++ src/internal/hashtable/insert.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "hashtable.h"
+#include "../../../include/minishell.h"
 
 void	hash_insert(t_ht_item *new_item, t_hashtable *table)
 {
